Phân loại rác/requirements_snapshot.txt
torch==2.8.0+cu128
torchvision==0.23.0+cu128
kagglehub==1.0.0
pandas==2.3.3
scikit-learn==1.7.2
Pillow==12.0.0
matplotlib==3.10.8
seaborn==0.13.2
PyYAML==6.0.3
notebook==7.5.5
jupyterlab==4.5.6
ipykernel==7.2.0
nbformat==5.10.4
